third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/inline-event-handler-blocked-after-injecting-meta-expected.txt

CONSOLE MESSAGE: line 8: Clicking a link, pre-policy:
CONSOLE MESSAGE: line 21: PASS: Event handler triggered pre-policy.
CONSOLE MESSAGE: line 14: Injecting Content-Security-Policy.
CONSOLE MESSAGE: line 19: Clicking a link, post-policy:
CONSOLE MESSAGE: line 21: FAIL: Event handler triggered post-policy.
This test checks that CSP is evaluated on each call to an inline event handler, even if it's been executed pre-policy. It passes if one 'PASS' and no 'FAIL' messages appear.
